+++ awssert/iam.py
@@ -2,11 +2,13 @@ from datetime import datetime
 
 import dateutil
 
-from awssert.prefixes import prefixes, AssertionPrefixes
-from awssert.core import AssertionPrefixRouter, BotoObjectProxyRegister, BotoObjectProxy
+from awssert.prefixes import prefixes
 
 
 class UserAssertions:
+
+    attaches_to = "iam.User"
+
     @prefixes(["has", "does_not_have"])
     def name(self, user, name):
         return name == user.name
@@ -22,11 +24,35 @@ class UserAssertions:
         return group in user.groups.all()
 
 
-def register_iam_assertions(class_attributes, base_classes, **kwargs):
-    proxy = BotoObjectProxy()
-    base_classes.insert(0, BotoObjectProxyRegister)
-    class_attributes["proxy"] = proxy
-    for prefix in AssertionPrefixes.all:
-        class_attributes[prefix] = AssertionPrefixRouter(
-            prefix, UserAssertions(), proxy
-        )
+class PolicyAssertions:
+
+    attaches_to = "iam.Policy"
+
+    @prefixes(["has", "does_not_have"])
+    def name(self, policy, name):
+        return name == policy.policy_name
+
+    @prefixes(["has", "does_not_have"])
+    def description(self, policy, description):
+        return description == policy.description
+
+    @prefixes(["was", "was_not"])
+    def created_at(self, policy, date):
+        if not isinstance(date, datetime):
+            date = dateutil.parser.parse(date)
+        return date == policy.create_date
+
+    @prefixes(["was", "was_not"])
+    def last_updated_at(self, policy, date):
+        if not isinstance(date, datetime):
+            date = dateutil.parser.parse(date)
+        return date == policy.update_date
+
+    @prefixes(["should_be", "should_not_be"])
+    def attached_to(self, policy, *entities):
+        attached = {
+            "iam.Group": policy.attached_groups.all(),
+            "iam.Role": policy.attached_roles.all(),
+            "iam.User": policy.attached_users.all(),
+        }
+        return all([entity in attached[type(entity).__name__] for entity in entities])